Updates! (Pre-reversion-backup)

+++ Value.hpp
@@ -76,7 +76,7 @@ namespace Jsonifier {
 		inline bool atFirstField() const noexcept;
 		inline void abandon() noexcept;
 		inline ValueIterator childValue() const noexcept;
-		inline int32_t depth() const noexcept;
+		inline size_t depth() const noexcept;
 		inline JsonifierResult<JsonType> type() const noexcept;
 		inline JsonifierResult<bool> startObject() noexcept;
 		inline JsonifierResult<bool> startRootObject() noexcept;
